End-of-line comments now work, as do nested-parsing multiple-statement lines. The source function now reports on what line a problem occurred and processing is appropriately terminated. Also fixed a few minor compilation problems in Visual Studio.

+++ src/core/datatypes/complex/container/Container.cpp
@@ -109,7 +109,7 @@ RbPtr<RbLanguageObject> Container::executeOperationSimple(const std::string& nam
         // get the member with give index
         RbPtr<const Natural> index( static_cast<const Natural*>( (const RbObject*)(*args)[0]->getValue()) );
 
-        if (size() < index->getValue()) {
+        if (size() < (size_t)(index->getValue()) ) {
             throw RbException("Index out of bounds in []");
         }
